@@ -2,24 +2,20 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
 
 import { ShoppingListComponent } from './shopping-list/shopping-list.component';
+import { HomeComponent } from './core/home/home.component';
 
-import { HomeComponent} from './core/home/home.component';
-import { AuthGuard } from './auth/auth-guard.service';
-
-const appRoutes:Routes=[
-	{ path :'', component: HomeComponent},
-	{ path:'recipes', loadChildren:'./recipes/recipes.module#RecipesModule'},	
-	{ path: 'shopping-list', component: ShoppingListComponent}
+const appRoutes: Routes = [
+  { path: '', component: HomeComponent },
+  { path: 'recipes', loadChildren: './recipes/recipes.module#RecipesModule'},
+  { path: 'shopping-list', component: ShoppingListComponent }
 ];
 
 @NgModule({
-	imports:[RouterModule.forRoot(appRoutes, {preloadingStrategy: PreloadAllModules})],
-	exports:[RouterModule],
-	providers:[AuthGuard]
-
+  imports: [
+    RouterModule.forRoot(appRoutes, {preloadingStrategy: PreloadAllModules})
+  ],
+  exports: [RouterModule]
 })
-
-
-export class AppRoutingModule{
+export class AppRoutingModule {
 
 }
