@@ -1,12 +1,16 @@
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Injectable } from '@angular/core';
-import { CanActivate,ActivatedRouteSnapshot,RouterStateSnapshot } from '@angular/router';
-import { AuthService } from './auth.service';
-
+import {Store} from '@ngrx/store';
+import * as fromApp from '../store/app.reducers';
+import * as fromAuth from './store/auth.reducers';
 @Injectable()
-export class AuthGuard implements CanActivate{
+export class AuthGuard implements CanActivate {
 
-	constructor(private authService:AuthService){}
-	canActivate(route:ActivatedRouteSnapshot, state: RouterStateSnapshot){
-		return this.authService.isAuthenticated();
-	}
+  constructor(private store:Store<fromApp.AppState>) {}
+
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
+    return this.store.select('auth')
+      .take(1)
+      .map((authState:fromAuth.State)=>{return authState.authenticated});
+  }
 }
