@@ -1,23 +1,18 @@
 import { NgModule } from '@angular/core';
-import { Routes,RouterModule} from '@angular/router';
-import { SigninComponent } from './signin/signin.component';
-import { SignupComponent } from './signup/signup.component';
+import { Routes, RouterModule } from '@angular/router';
 
-const authRoutes :Routes=[
-	{ path: 'signup', component: SignupComponent},
-	{ path: 'signin', component: SigninComponent}
-]
+import { SignupComponent } from './signup/signup.component';
+import { SigninComponent } from './signin/signin.component';
+
+const authRoutes: Routes = [
+  { path: 'signup', component: SignupComponent },
+  { path: 'signin', component: SigninComponent },
+];
 
 @NgModule({
-	
-	imports:[
-		RouterModule.forChild(authRoutes)
-	],
-	exports:[
-		RouterModule
-	]
-
+  imports: [
+    RouterModule.forChild(authRoutes)
+  ],
+  exports: [RouterModule]
 })
-export class AuthRoutingModule{
-	
-}
+export class AuthRoutingModule {}
