@@ -1,6 +1,9 @@
-import { Component, OnInit } from '@angular/core';
-import { NgForm } from '@angular/forms';
-import { AuthService} from '../auth.service';
+import {Component, OnInit} from '@angular/core';
+import {NgForm} from '@angular/forms';
+
+import * as fromApp from '../../store/app.reducers';
+import * as AuthActions from '../store/auth.actions';
+import {Store} from "@ngrx/store";
 
 @Component({
   selector: 'app-signin',
@@ -9,16 +12,16 @@ import { AuthService} from '../auth.service';
 })
 export class SigninComponent implements OnInit {
 
-  constructor(private authService: AuthService) { }
+  constructor(private store: Store<fromApp.AppState>) {
+  }
 
   ngOnInit() {
   }
 
-  onSignin(form:NgForm){
-  	const email = form.value.email;
-  	const password = form.value.password;
-  	this.authService.signinUser(email,password);
-  	
+  onSignin(form: NgForm) {
+    const email = form.value.email;
+    const password = form.value.password;
+    this.store.dispatch(new AuthActions.TrySignin({username: email, password: password}))
   }
 
 }
