@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
-import { AuthService } from '../auth.service';
+import {Store} from '@ngrx/store';
+import * as fromApp from '../../store/app.reducers';
+import * as AuthActions from '../store/auth.actions';
+
 @Component({
   selector: 'app-signup',
   templateUrl: './signup.component.html',
@@ -8,15 +11,15 @@ import { AuthService } from '../auth.service';
 })
 export class SignupComponent implements OnInit {
 
-  constructor(private authService:AuthService) { }
+  constructor(private store:Store<fromApp.AppState>) { }
 
   ngOnInit() {
   }
 
-  onSignup(form : NgForm){
-  	const email = form.value.email;
-  	const password = form.value.password;	
-  	this.authService.signupUser(email,password);
+  onSignup(form: NgForm) {
+    const email = form.value.email;
+    const password = form.value.password;
+    this.store.dispatch(new AuthActions.TrySignup({username:email, password:password}));
   }
 
 }
