@@ -1,28 +1,33 @@
 import { NgModule } from '@angular/core';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { HeaderComponent } from './header/header.component';
 import { HomeComponent } from './home/home.component';
 import { SharedModule } from '../shared/shared.module';
 import { AppRoutingModule } from '../app-routing.module';
-
-import { ShoppingListService } from '../shopping-list/shopping-list.service';
-import { RecipeService } from '../recipes/recipe.service';
 import { DataStorageService } from '../shared/data-storage.service';
-import { AuthService } from '../auth/auth.service';
+import { RecipeService } from '../recipes/recipe.service';
+import { AuthInterceptor } from '../shared/auth.interceptor';
+import { LoggingInterceptor } from '../shared/logging.interceptor';
 
 @NgModule({
-	declarations:[
-		HeaderComponent,
-		HomeComponent
-	],
-	imports:[
-		SharedModule,
-		AppRoutingModule
-	],
-	exports:[AppRoutingModule, HeaderComponent],
-	providers: [ShoppingListService, RecipeService, DataStorageService, AuthService],
-
+  declarations: [
+    HeaderComponent,
+    HomeComponent
+  ],
+  imports: [
+    SharedModule,
+    AppRoutingModule
+  ],
+  exports: [
+    AppRoutingModule,
+    HeaderComponent
+  ],
+  providers: [
+    RecipeService,
+    DataStorageService,
+    {provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true},
+    {provide: HTTP_INTERCEPTORS, useClass: LoggingInterceptor, multi: true}
+  ]
 })
-export class CoreModule{
-	
-}
+export class CoreModule {}
