@@ -1,34 +1,38 @@
-import { Component} from '@angular/core';
-import { Response } from '@angular/http';
+import { Component, OnInit } from '@angular/core';
+import {Store} from '@ngrx/store';
+import * as fromApp from '../../store/app.reducers';
+import * as fromAuth from '../../auth/store/auth.reducers';
 import { DataStorageService } from '../../shared/data-storage.service';
-import { AuthService } from '../../auth/auth.service';
+import {Observable} from "rxjs/Observable";
+import * as AuthActions from '../../auth/store/auth.actions';
+
 @Component({
-  selector:'app-header',
-  templateUrl: './header.component.html',
-  styleUrls:['./header.component.css']
+  selector: 'app-header',
+  templateUrl: './header.component.html'
 })
-export class HeaderComponent{
-	constructor( private dsService: DataStorageService
-		,private authService: AuthService){}
-	onSave(){
-		this.dsService.storeRecipes().subscribe(
-			(response:Response) => {
-				console.log(response);
-			},
-			(error:Response) => {
+export class HeaderComponent implements OnInit{
+  authState: Observable<fromAuth.State>;
+  ngOnInit(){
+    this.authState = this.store.select('auth');
+  }
 
-			}
-		);
-	}
+  constructor(private dataStorageService: DataStorageService,private store:Store<fromApp.AppState>) {
+  }
 
-	onLogout(){
-		this.authService.logout();
-	}
-	onFetch(){
-		this.dsService.getRecipes();
-	}
+  onSaveData() {
+    this.dataStorageService.storeRecipes()
+      .subscribe(
+        (response) => {
+          console.log(response);
+        }
+      );
+  }
 
-	isAuthenticated() {
-    return this.authService.isAuthenticated();
+  onFetchData() {
+    this.dataStorageService.getRecipes();
+  }
+
+  onLogout() {
+    this.store.dispatch(new AuthActions.Logout());
   }
 }
