@@ -1,46 +1,48 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute,Router, Params } from '@angular/router';
+import { ActivatedRoute, Params, Router } from '@angular/router';
 
 import { Recipe } from '../recipe.model';
 import { RecipeService } from '../recipe.service';
-import { Ingredient } from '../../shared/ingredient.model';
-import { ShoppingListService } from '../../shopping-list/shopping-list.service';
-
+import { Store} from '@ngrx/store';
+import * as ShoppingListActions from "../../shopping-list/store/shopping-list.actions";
+import * as fromApp from '../../store/app.reducers';
 @Component({
   selector: 'app-recipe-detail',
   templateUrl: './recipe-detail.component.html',
   styleUrls: ['./recipe-detail.component.css']
 })
 export class RecipeDetailComponent implements OnInit {
-  
-  recipe:Recipe;
+  recipe: Recipe;
   id: number;
-  constructor(private shoppingListService:ShoppingListService, 
-    private router:Router,
-    private route:ActivatedRoute, private recipeService:RecipeService ) { }
+
+  constructor(private recipeService: RecipeService,
+              private route: ActivatedRoute,
+              private router: Router, private store:Store<fromApp.AppState>) {
+  }
 
   ngOnInit() {
-    //this.id = this.route.snapshot.params["id"];
-    this.route.params.subscribe(
-      (params:Params) => {
-        this.id = +params["id"];
-        this.recipe = this.recipeService.getRecipe(this.id);
-      }
+    this.route.params
+      .subscribe(
+        (params: Params) => {
+          this.id = +params['id'];
+          this.recipe = this.recipeService.getRecipe(this.id);
+        }
       );
-
   }
 
-  addToShoppingList(ingredients: Ingredient[]){
-  	this.shoppingListService.addIngredients(ingredients);
+  onAddToShoppingList() {
+    //this.recipeService.addIngredientsToShoppingList(this.recipe.ingredients);
+    this.store.dispatch(new ShoppingListActions.AddIngredients(this.recipe.ingredients));
   }
 
-  onEditRecipe(){
-    // this.router.navigate(['edit'], {relativeTo: this.route});
-    this.router.navigate(['../', this.id, 'edit'], {relativeTo: this.route});
+  onEditRecipe() {
+    this.router.navigate(['edit'], {relativeTo: this.route});
+    // this.router.navigate(['../', this.id, 'edit'], {relativeTo: this.route});
   }
 
-  onDeleteRecipe(){
+  onDeleteRecipe() {
     this.recipeService.deleteRecipe(this.id);
     this.router.navigate(['/recipes']);
   }
+
 }
