@@ -1,93 +1,105 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute,Params,Router } from '@angular/router';
+import { ActivatedRoute, Params, Router } from '@angular/router';
 import { FormGroup, FormControl, FormArray, Validators } from '@angular/forms';
+
 import { RecipeService } from '../recipe.service';
-import { Recipe } from '../recipe.model';
+
 @Component({
   selector: 'app-recipe-edit',
   templateUrl: './recipe-edit.component.html',
   styleUrls: ['./recipe-edit.component.css']
 })
 export class RecipeEditComponent implements OnInit {
-  id:number;
+  id: number;
   editMode = false;
-  recipeForm:FormGroup;
-  constructor(private route: ActivatedRoute, private recipeService:RecipeService,
-    private router:Router) { }
+  recipeForm: FormGroup;
 
-  ngOnInit() {
-  	this.route.params.subscribe(
-  		(params:Params) => {
-  			this.id = +params['id'];
-  			this.editMode = params['id'] !=null;
-        this.initForm();
-  		}
-  		);
+  constructor(private route: ActivatedRoute,
+              private recipeService: RecipeService,
+              private router: Router) {
   }
 
-  private initForm(){
-    let recipeName= '';
+  ngOnInit() {
+    this.route.params
+      .subscribe(
+        (params: Params) => {
+          this.id = +params['id'];
+          this.editMode = params['id'] != null;
+          this.initForm();
+        }
+      );
+  }
+
+  onSubmit() {
+    // const newRecipe = new Recipe(
+    //   this.recipeForm.value['name'],
+    //   this.recipeForm.value['description'],
+    //   this.recipeForm.value['imagePath'],
+    //   this.recipeForm.value['ingredients']);
+    if (this.editMode) {
+      this.recipeService.updateRecipe(this.id, this.recipeForm.value);
+    } else {
+      this.recipeService.addRecipe(this.recipeForm.value);
+    }
+    this.onCancel();
+  }
+
+  onAddIngredient() {
+    (<FormArray>this.recipeForm.get('ingredients')).push(
+      new FormGroup({
+        'name': new FormControl(null, Validators.required),
+        'amount': new FormControl(null, [
+          Validators.required,
+          Validators.pattern(/^[1-9]+[0-9]*$/)
+        ])
+      })
+    );
+  }
+
+  onDeleteIngredient(index: number) {
+    (<FormArray>this.recipeForm.get('ingredients')).removeAt(index);
+  }
+
+  onCancel() {
+    this.router.navigate(['../'], {relativeTo: this.route});
+  }
+
+  private initForm() {
+    let recipeName = '';
     let recipeImagePath = '';
     let recipeDescription = '';
     let recipeIngredients = new FormArray([]);
-    if(this.editMode){
+
+    if (this.editMode) {
       const recipe = this.recipeService.getRecipe(this.id);
       recipeName = recipe.name;
       recipeImagePath = recipe.imagePath;
       recipeDescription = recipe.description;
-      if(recipe['ingredients']){
-        for(let ing of recipe.ingredients){
+      if (recipe['ingredients']) {
+        for (let ingredient of recipe.ingredients) {
           recipeIngredients.push(
-              new FormGroup({
-                 'name': new FormControl(ing.name, Validators.required),
-                 'amount': new FormControl(ing.amount, [Validators.required, Validators.pattern(/^[1-9]+[0-9]*$/)])
-              })
-           );
+            new FormGroup({
+              'name': new FormControl(ingredient.name, Validators.required),
+              'amount': new FormControl(ingredient.amount, [
+                Validators.required,
+                Validators.pattern(/^[1-9]+[0-9]*$/)
+              ])
+            })
+          );
         }
       }
     }
 
     this.recipeForm = new FormGroup({
-      'name' : new FormControl(recipeName, Validators.required),
-      'imagePath' : new FormControl(recipeImagePath, Validators.required),
-      'description' : new FormControl(recipeDescription, Validators.required),
+      'name': new FormControl(recipeName, Validators.required),
+      'imagePath': new FormControl(recipeImagePath, Validators.required),
+      'description': new FormControl(recipeDescription, Validators.required),
       'ingredients': recipeIngredients
     });
-  }
-  onAddIngredients(){
-
-    (<FormArray>this.recipeForm.get('ingredients')).push(
-        new FormGroup({
-          'name' : new FormControl(null, Validators.required),
-          'amount' : new FormControl(null,  [Validators.required, Validators.pattern(/^[1-9]+[0-9]*$/)])
-
-          })
-      );
-  }
-
-  onDeleteIngredients(index:number){
-    (<FormArray>this.recipeForm.get('ingredients')).removeAt(index);
-  }
-
-  onSubmit(){
-    /*const newRecipe = new Recipe(this.recipeForm.value['name'], 
-      this.recipeForm.value['description'],
-      this.recipeForm.value['imagePath'],
-      this.recipeForm.value['ingredients']);*/
-    if(this.editMode){
-      this.recipeService.updateRecipe(this.id, this.recipeForm.value );
-    }else{
-      this.recipeService.addRecipe(this.recipeForm.value);
-    }
-
-    this.router.navigate(['../'], {relativeTo:this.route});
   }
 
   getControls() {
     return (<FormArray>this.recipeForm.get('ingredients')).controls;
   }
 
-  onCancel(){
-    this.router.navigate(['../'], {relativeTo:this.route});
-  }
 }
