@@ -1,13 +1,16 @@
 import { Component, OnInit, Input } from '@angular/core';
+
 import { Recipe } from '../../recipe.model';
+
 @Component({
   selector: 'app-recipe-item',
   templateUrl: './recipe-item.component.html',
   styleUrls: ['./recipe-item.component.css']
 })
 export class RecipeItemComponent implements OnInit {
-  @Input() recipeItem:Recipe;
-  @Input() index:number;
+  @Input() recipe: Recipe;
+  @Input() index: number;
+
   ngOnInit() {
   }
 }
