@@ -1,31 +1,39 @@
-import { Component, OnInit } from '@angular/core';
-
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
+import { Subscription } from 'rxjs/Subscription';
+
 import { Recipe } from '../recipe.model';
 import { RecipeService } from '../recipe.service';
+
 @Component({
   selector: 'app-recipe-list',
   templateUrl: './recipe-list.component.html',
   styleUrls: ['./recipe-list.component.css']
 })
-export class RecipeListComponent implements OnInit {
-  recipes:Recipe[];
-  constructor(private recipeService: RecipeService, 
-  	private router:Router, private route: ActivatedRoute) { }
+export class RecipeListComponent implements OnInit, OnDestroy {
+  recipes: Recipe[];
+  subscription: Subscription;
+
+  constructor(private recipeService: RecipeService,
+              private router: Router,
+              private route: ActivatedRoute) {
+  }
 
   ngOnInit() {
-    this.recipeService.recipesChange.subscribe(
-      (recipesB:Recipe[]) => {
-        this.recipes = recipesB;
-      }
+    this.subscription = this.recipeService.recipesChanged
+      .subscribe(
+        (recipes: Recipe[]) => {
+          this.recipes = recipes;
+        }
       );
-
     this.recipes = this.recipeService.getRecipes();
   }
 
-  onNewRecipe(){
-  	this.router.navigate(['new'], { relativeTo: this.route });
+  onNewRecipe() {
+    this.router.navigate(['new'], {relativeTo: this.route});
   }
- 
 
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
+  }
 }
