@@ -1,59 +1,54 @@
-import { Injectable,OnInit } from '@angular/core';
+import { Subject } from 'rxjs/Subject';
 import { Recipe } from './recipe.model';
 import { Ingredient } from '../shared/ingredient.model';
-import { Subject } from 'rxjs/Subject';
-import { ShoppingListService } from '../shopping-list/shopping-list.service';
+export class RecipeService {
+  recipesChanged = new Subject<Recipe[]>();
 
-@Injectable()
-export class RecipeService{
-  constructor(private slService: ShoppingListService ){
+  private recipes: Recipe[] = [
+    new Recipe(
+      'Tasty Schnitzel',
+      'A super-tasty Schnitzel - just awesome!',
+      'https://upload.wikimedia.org/wikipedia/commons/7/72/Schnitzel.JPG',
+      [
+        new Ingredient('Meat', 1),
+        new Ingredient('French Fries', 20)
+      ]),
+    new Recipe('Big Fat Burger',
+      'What else you need to say?',
+      'https://upload.wikimedia.org/wikipedia/commons/b/be/Burger_King_Angus_Bacon_%26_Cheese_Steak_Burger.jpg',
+      [
+        new Ingredient('Buns', 2),
+        new Ingredient('Meat', 1)
+      ])
+  ];
 
-  }
+  constructor() {}
 
-
-
-  recipesChange = new Subject<Recipe[]>();
-	private recipes: Recipe[] = [
-    new Recipe('Pizza', 'Taste good!', 'https://cdn.modpizza.com/wp-content/uploads/2016/11/mod-pizza-maddy-default-e1479167621575.png', [
-    		new Ingredient("Pepperoni", 3),
-    		new Ingredient("Cheese", 4)
-
-    	]),
-    new Recipe('A Test Recipe', 'This is simply a test', 'https://upload.wikimedia.org/wikipedia/commons/1/15/Recipe_logo.jpeg',[
-    		new Ingredient("buns", 2),
-    		new Ingredient("meat", 4)
-    	])
-    ];
-
-  setRecipes(recipes: Recipe[]){
+  setRecipes(recipes: Recipe[]) {
     this.recipes = recipes;
-    this.recipesChange.next(this.recipes);
+    this.recipesChanged.next(this.recipes.slice());
   }
 
-  getRecipes(){
-  	return this.recipes.slice();
+  getRecipes() {
+    return this.recipes.slice();
   }
 
-  getRecipe(id:number){
-    return this.recipes[id];
+  getRecipe(index: number) {
+    return this.recipes[index];
   }
 
-  addRecipe(recipe:Recipe){
+  addRecipe(recipe: Recipe) {
     this.recipes.push(recipe);
-    this.slService.addIngredients(recipe.ingredients);
-    this.recipesChange.next(this.recipes.slice());
+    this.recipesChanged.next(this.recipes.slice());
   }
 
-  updateRecipe(index:number, newRecipe:Recipe){
+  updateRecipe(index: number, newRecipe: Recipe) {
     this.recipes[index] = newRecipe;
-    this.recipesChange.next(this.recipes.slice());
+    this.recipesChanged.next(this.recipes.slice());
   }
 
-  deleteRecipe(index:number){
-    this.recipes.splice(index,1);
-    this.recipesChange.next(this.recipes.slice());
-
-    
-
+  deleteRecipe(index: number) {
+    this.recipes.splice(index, 1);
+    this.recipesChanged.next(this.recipes.slice());
   }
 }
