@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
-
+import { CommonModule } from '@angular/common';
 
 import { RecipesComponent } from './recipes.component';
 import { RecipeStartComponent } from './recipe-start/recipe-start.component';
@@ -9,25 +8,23 @@ import { RecipeListComponent } from './recipe-list/recipe-list.component';
 import { RecipeEditComponent } from './recipe-edit/recipe-edit.component';
 import { RecipeDetailComponent } from './recipe-detail/recipe-detail.component';
 import { RecipeItemComponent } from './recipe-list/recipe-item/recipe-item.component';
-import { RecipesRoutingModule} from './recipes-routing.module';
-import { SharedModule} from '../shared/shared.module';
+import { RecipesRoutingModule } from './recipes-routing.module';
+import { SharedModule } from '../shared/shared.module';
+
 @NgModule({
-	declarations:[
-		RecipesComponent,
-		RecipeStartComponent,
-		RecipeListComponent,
-		RecipeEditComponent,
-		RecipeDetailComponent,
-		RecipeItemComponent,
-	],
-	imports:[
-		CommonModule,
-    	ReactiveFormsModule,
-    	RecipesRoutingModule,
-    	SharedModule
-    ]
+  declarations: [
+    RecipesComponent,
+    RecipeStartComponent,
+    RecipeListComponent,
+    RecipeEditComponent,
+    RecipeDetailComponent,
+    RecipeItemComponent
+  ],
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    RecipesRoutingModule,
+    SharedModule
+  ]
 })
-export class RecipesModule{
-
-
-}
+export class RecipesModule {}
