@@ -1,42 +1,49 @@
-import {Injectable} from '@angular/core';
-import { Http, Response} from '@angular/http';
+import { Injectable } from '@angular/core';
+import { HttpClient, HttpHeaders, HttpParams, HttpRequest } from '@angular/common/http';
+import 'rxjs/Rx';
+
 import { RecipeService } from '../recipes/recipe.service';
 import { Recipe } from '../recipes/recipe.model';
-import { AuthService } from '../auth/auth.service';
-import 'rxjs/Rx';
+
 @Injectable()
-export class DataStorageService{
-	constructor(private http:Http, private recipeService:RecipeService,
-		private authService: AuthService){}
+export class DataStorageService {
+  constructor(private httpClient: HttpClient,
+              private recipeService: RecipeService) {
+  }
 
-	storeRecipes(){
-		const token = this.authService.getToken();
-		return this.http.put('https://ng-recipe-book-5849b.firebaseio.com/recipes.json?auth=' + token, this.recipeService.getRecipes());
-	}
+  storeRecipes() {
+    // const headers = new HttpHeaders().set('Authorization', 'Bearer afdklasflaldf');
 
-	getRecipes(){
-		const token = this.authService.getToken();
-		
-		return this.http.get('https://ng-recipe-book-5849b.firebaseio.com/recipes.json?auth='+ token)
-		.map(
-			(response:Response) =>{
-				const recipes: Recipe[] = response.json();
-				for(let recipe of recipes){
-					if(!recipe['ingredients']){
-						recipe['ingredients'] = [];
-					}
-				}
-			return recipes;
-			}
-		)
-		.subscribe(
-			(recipes:Recipe[]) => {
-				
-				this.recipeService.setRecipes(recipes);
-				console.log(recipes);
-			}
-		);
-	}
+    // return this.httpClient.put('https://ng-recipe-book-3adbb.firebaseio.com/recipes.json', this.recipeService.getRecipes(), {
+    //   observe: 'body',
+    //   params: new HttpParams().set('auth', token)
+    //   // headers: headers
+    // });
+    const req = new HttpRequest('PUT', 'https://ng-rx-3c367.firebaseio.com//recipes.json', this.recipeService.getRecipes(), {reportProgress: true});
+    return this.httpClient.request(req);
+  }
 
-
+  getRecipes() {
+    // this.httpClient.get<Recipe[]>('https://ng-recipe-book-3adbb.firebaseio.com/recipes.json?auth=' + token)
+    this.httpClient.get<Recipe[]>('https://ng-rx-3c367.firebaseio.com//recipes.json', {
+      observe: 'body',
+      responseType: 'json'
+    })
+      .map(
+        (recipes) => {
+          console.log(recipes);
+          for (let recipe of recipes) {
+            if (!recipe['ingredients']) {
+              recipe['ingredients'] = [];
+            }
+          }
+          return recipes;
+        }
+      )
+      .subscribe(
+        (recipes: Recipe[]) => {
+          this.recipeService.setRecipes(recipes);
+        }
+      );
+  }
 }
