@@ -1,21 +1,12 @@
-import { Directive, OnInit, HostListener,HostBinding } from '@angular/core';
+import { Directive, HostListener, HostBinding } from '@angular/core';
 
 @Directive({
-	selector:'[appDropdown]'
-
+  selector: '[appDropdown]'
 })
-export class DropdownDirective implements OnInit{
-	@HostBinding('class.open') isOpen = false;
+export class DropdownDirective {
+  @HostBinding('class.open') isOpen = false;
 
-	@HostListener('click') toggleOpen (eventData: Event){
-		this.isOpen = !this.isOpen;
-	}
-
-	constructor(){
-
-	}
-
-	ngOnInit(){
-
-	}
+  @HostListener('click') toggleOpen() {
+    this.isOpen = !this.isOpen;
+  }
 }
