@@ -1,4 +1,3 @@
-export class Ingredient{
-  constructor(public name:string, public amount:number){
-  }
+export class Ingredient {
+  constructor(public name: string, public amount: number) {}
 }
