@@ -1,8 +1,15 @@
-import { Component, OnInit, OnDestroy, ViewChild} from '@angular/core';
+import {
+  Component,
+  OnInit,
+  OnDestroy,
+  ViewChild
+} from '@angular/core';
 import { NgForm } from '@angular/forms';
-import { Ingredient } from '../../shared/ingredient.model';
-import { ShoppingListService } from '../shopping-list.service';
 import { Subscription } from 'rxjs/Subscription';
+import { Store} from '@ngrx/store';
+import { Ingredient } from '../../shared/ingredient.model';
+import * as ShoppingListActions from '../store/shopping-list.actions';
+import * as fromApp from '../../store/app.reducers';
 
 @Component({
   selector: 'app-shopping-edit',
@@ -10,53 +17,56 @@ import { Subscription } from 'rxjs/Subscription';
   styleUrls: ['./shopping-edit.component.css']
 })
 export class ShoppingEditComponent implements OnInit, OnDestroy {
-
-  @ViewChild('f') shoppingListForm : NgForm;
-  subscription:Subscription;
+  @ViewChild('f') slForm: NgForm;
+  subscription: Subscription;
   editMode = false;
-  editIndex :number;
-  editItem :Ingredient;
+  editedItem: Ingredient;
 
-  constructor(private shoppinglistService: ShoppingListService) { }
+  constructor( private store:Store<fromApp.AppState>) { }
 
   ngOnInit() {
-    this.subscription = this.shoppinglistService.startEditing.subscribe(
-      (i:number)=>{
-        this.editIndex = i;
-        this.editMode = true;
-        this.editItem = this.shoppinglistService.getIngredient(i);
-        this.shoppingListForm.setValue({
-          name: this.editItem.name,
-          amount : this.editItem.amount
-        })
-      });
+    this.subscription= this.store.select('shoppingList').subscribe(
+      data => {
+        if(data.editedIngredientIndex >-1){
+          this.editedItem= data.editedIngredient;
+          this.editMode = true;
+          this.slForm.setValue({
+            name: this.editedItem.name,
+            amount: this.editedItem.amount
+          })
+        }else{
+          this.editMode = false;
+        }
+      }
+    );
   }
-  
-  ngOnDestroy(){
+
+  onSubmit(form: NgForm) {
+    const value = form.value;
+    const newIngredient = new Ingredient(value.name, value.amount);
+    if (this.editMode) {
+      this.store.dispatch(new ShoppingListActions.UpdateIngredient({ingredient:newIngredient}));
+    } else {
+      this.store.dispatch(new ShoppingListActions.AddIngredient(newIngredient));
+      //this.slService.addIngredient(newIngredient);
+    }
+    this.editMode = false;
+    form.reset();
+  }
+
+  onClear() {
+    this.slForm.reset();
+    this.editMode = false;
+  }
+
+  onDelete() {
+    this.store.dispatch(new ShoppingListActions.DeleteIngredient());
+    this.onClear();
+  }
+
+  ngOnDestroy() {
+    this.store.dispatch(new ShoppingListActions.StopEdit());
     this.subscription.unsubscribe();
   }
 
-  onAddItem(form: NgForm){
-    const value = form.value;
-  	const ing = new Ingredient(value.name, value.amount);
-    if(this.editMode){
-      this.shoppinglistService.updateIngredient(this.editIndex, ing);
-    }else{
-      this.shoppinglistService.addIngredient(ing);
-    }
-    this.shoppingListForm.reset();
-    this.editMode = false;
-
-  }
-
-  onDelete(){
-    this.onClear();
-    this.shoppinglistService.deleteIngredient(this.editIndex);
-  }
-
-  onClear(){
-    this.shoppingListForm.reset();
-    this.editMode = false;
-  }
 }
-
