@@ -1,31 +1,26 @@
-import { Component, OnInit,OnDestroy } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import{Store} from '@ngrx/store';
 import { Ingredient } from '../shared/ingredient.model';
-import { ShoppingListService } from './shopping-list.service';
-import { Subscription } from 'rxjs/Subscription';
+import {Observable} from "rxjs/Observable";
+import * as ShoppingListActions from './store/shopping-list.actions';
+import * as fromApp from '../store/app.reducers';
 
 @Component({
   selector: 'app-shopping-list',
   templateUrl: './shopping-list.component.html',
   styleUrls: ['./shopping-list.component.css']
 })
-export class ShoppingListComponent implements OnInit, OnDestroy{
-  ingredients: Ingredient[];
-  private subscription :Subscription;
-  constructor(private shoppinglistService: ShoppingListService) { }
+export class ShoppingListComponent implements OnInit {
+  shoppingListState: Observable<{ingredients:Ingredient[]}>;
+
+  constructor( private store:Store<fromApp.AppState>) { }
 
   ngOnInit() {
-    this.ingredients = this.shoppinglistService.getIngredients();
-    this.subscription = this.shoppinglistService.ingredientChanged.subscribe(
-        (ingredients: Ingredient[]) => {
-          this.ingredients = ingredients;
-        }
-      );
-  }
-  ngOnDestroy(){
-    this.subscription.unsubscribe();
+    this.shoppingListState = this.store.select('shoppingList');
   }
 
-  onEditItem(i:number){
-    this.shoppinglistService.startEditing.next(i);
+  onEditItem(index: number) {
+    this.store.dispatch(new ShoppingListActions.StartEdit(index));
   }
+
 }
